requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=1.5.3
openai>=1.3.0
numpy>=1.24.3